requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
pandas==2.2.1
openpyxl==3.1.2
gunicorn==21.2.0
Flask-Login==0.6.3
Werkzeug==3.0.1
Flask-Mail==0.9.1
itsdangerous==2.1.2 